Add iterable dataset
Should use a lot less RAM, as loading att ~144 sets of images grows to more than 30GB of ram at once. Is pushing me to more go towards Zarr probably, but one last try for webdataset for this thing. A lot more limitations with this one compared to the other SatFlowDataset, but tradeoffs it for memory