Update training/test/val splits based off #28 critique
The correlation between days of the week might mean that there is leakage between the training and val/test sets. This is now split up so that 2020 is used for training, and 2021 is used for val/testing. This does mean that there is only data for the first half of 2021 in the test/val set for now, but that does cover winter/spring/part of summer so far, with more as time goes on.