Add training step for NowcastingGAN and losses
There is the issue that I'm not entirely sure of, in the paper they say they use 6 samples of the latent space per input and use that for the mean prediction for the grid cell regularizer. It sounds like they also do that for the generator and discriminator losses, so I think I am doing it right by getting the mean prediction from 6 calls to the generator, but that could be wrong.